Adds TabPanel Component
Adds a TabPanel Component. Previously Tab was being split into a
TabButton and then the rest of the tab was being rendered. This worked
for the most part. However, Tab needs to act more as a virtual parent
container to tabButton and its tabPanel instead of acting as both
the parent and one of the children. It was leading to some weird bugs
with props being set and the componentWillReceiveProps method.

To solve this issue, TabPanel is introduced to take the child role that
Tab was playing and now Tab is solely the parent. The render method
of Tab now clearly shows that it should never be used for rendering.

Renames some methods, states, and props from collapsibles to
collapsiblesGroups in order to be more clear.

Removes cx as an unused variable.